added arena allocator

+++ template.c
@@ -7,6 +7,6 @@
 #include "scimath.h"
 #include "template.h"
 
-ksm_GENERIC_VECTOR_IMPL(int)
+ksm_GENERIC_VECTOR_IMPL(size_t)
 ksm_GENERIC_VECTOR_IMPL(double)
 ksm_GENERIC_VECTOR_IMPL(char)
